Calculates shipping turnaround time
Uses a table generated with a SQL Make Table query. The file generated contains a list of all items ordered between 2/23/19 and 3/25/19, the last 30 days of data in the supplied dataset. It gives the amount of time between when the order was made and when the order was shipped, for each order that has shipped. Orders that hadn't been shipped yet get no turnaround time assigned to them and are not included in the analysis.